requirements.txt
pytest==7.2.1
selenium==4.8.2
webdriver-manager==3.8.5
faker==17.0.0
allure-python-commons-test==2.12.0
allure-python-commons==2.12.0
allure-pytest==2.12.0
allure-behave==2.12.0